requirements.txt
ib_insync>=0.9.86
pandas>=2.0.0
numpy>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
APScheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
flake8>=7.0.0
mypy>=1.8.0
